+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## FUTURE
 
 * [ADDED] - optional `context` object argument to the `ServiceError.prototype.buildFrom` method
+* [FIXED] - `Route` error handler `.then()` clause has been creating "runaway" promise
+* [FIXED] - errors which are not instanceof `Error` are not supported. The case of eg. `throw undefined` has not been properly handled.
 
 ## v1.0.2